Add notes on how documentation, decisions are made
Within any given structure or organisational process, decisions need to
be made collaboratively, approved and integrated into standard process.

This somewhat informal process mirrors the much stricter process of
managing logic changes within software. Accordingly, it stands to reason
that business decisions can be expressed in terms of the same version
control as software, if the "software" that is being modified is
documentation.

This commit adds documentation that, in turn, proposes itself.

== Design Notes ==

=== Asciidoc ===

The documentation is required to be written in the otherwise bespoke
format of "AsciiDoc". This format has several significant downsides:

- It's not in widespread use
- Support is limited
- It's a bit weird

However, the author is currently experimenting with the format and is
currently finding the positives of the language offset the negatives:

- It allows expressing documentation in a number of ways, such as a
  website, book etc.
- It is not dissimilar to the Markdown syntax more commonly used
- It is extensible and the more fuller featured language is useful in
  conveying longform ideas.
- Its in the same format that O'Reilly uses, and the author at some
  point wants to write a book. So, it's good practice.

Accordingly, this format is chosen.