Add notes about business lifetime
There are several strategies that surround business development. They
include:

Profitable, long term investment:

  Should a company be designed to yield a long term return on investment
  the company needs to make decisions that will pay out indefinitely,
  up to and including the 5 year forseable future.

Leveraging out for sale:

  Should a company be designed for sale, it's primary value is in the
  trust it builds for users. While this is still valuable, the goal is
  ultimately to sell those users out to a third party company, that may
  or may not retain the software.

In this case the author feels that the "long, slow burn" of a long term
company is infinitely more valuable than leveraging out for sale; even
if leveraging out is more profitable.

The financial gain is offset by the lost of trust in the community of
colleagues and users.