Remove the website
The initial design of this repository was to have a monolithic
repository, in the style ascribed by Google. However, the more recent
thinking the author has had is that each service that is deployed
separate should have its own separate git repository.

This allows a nice balance between reusability (by vendoring deployment
solutions) and independence.

This commit moves is one half of that required to move the website to:

  https://github.com/moarfit/www.moar.fit