Shift the business documentation to allow expansion
Currently there is a limited set of "business" documentation which
provides the outline for the moar.fit business, designed to develop and
serve the moar.fit product.

However, a business is an inherently complex structure that requires
more than the overview that's been currently supplied.

This commit moves the business document into its own folder, and renames
it "overview". Future work will start to establish further documentation
about the nature of the business, it's partners, activities, revenue
streams etc.